better data management and view modes
